requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
azure-core==1.36.0
beautifulsoup4==4.14.3
blinker==1.9.0
botbuilder-core==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
Flask==3.1.2
frozenlist==1.8.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpickle==1.4.2
MarkupSafe==3.0.3
msal==1.34.0
msrest==0.7.1
multidict==6.7.0
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
yarl==1.22.0
